feat(pulse-fetch): add dual-mode integration testing for source and built output
- Create test runner framework that runs tests in both SOURCE and BUILT modes
- SOURCE mode tests against TypeScript build output
- BUILT mode simulates exact npm publish process to test packaged output
- Restructure tests into separate files: source, built, and main integration
- Add comprehensive README documenting the dual-mode testing approach
- Update vitest config with proper timeouts and test isolation
- Add .test-publish directory to gitignore for built mode artifacts

This ensures the published npm package works exactly as tested, catching
packaging issues before they reach production.
