+++ productionized/pulse-fetch/package.json
@@ -19,7 +19,9 @@
     "test": "vitest",
     "test:ui": "vitest --ui",
     "test:run": "vitest run",
-    "test:integration": "npm run build:test && vitest run -c vitest.config.integration.ts && npm run clean",
+    "test:integration": "npm run build:test && vitest run -c vitest.config.integration.ts tests/integration/pulse-fetch.integration.test.ts && npm run clean",
+    "test:integration:source": "npm run build:test && vitest run -c vitest.config.integration.ts tests/integration/pulse-fetch.source.integration.test.ts && npm run clean",
+    "test:integration:built": "npm run build:test && vitest run -c vitest.config.integration.ts tests/integration/pulse-fetch.built.integration.test.ts && npm run clean",
     "test:integration:watch": "vitest -c vitest.config.integration.ts",
     "test:manual": "vitest run -c vitest.config.manual.ts",
     "test:all": "npm run test:run && npm run test:integration",
