+++ productionized/pulse-fetch/shared/package.json
@@ -25,7 +25,7 @@
     "@types/jsdom": "^21.1.7",
     "@types/node": "^24.0.0",
     "typescript": "^5.8.3",
-    "vitest": "^2.1.8"
+    "vitest": "^3.2.3"
   },
   "keywords": [],
   "author": "PulseMCP",
